Remove older files
